If there is no theme, use a default color for loading bar.
That is the one-line change in overlay.js.

While trying to figure out what was going on, I remodularized some code, putting
some things where they "belong". There is no functional change with these changes,
they are purely cosmetic.

- The code for actually loading the point clouds should not be part of
  calibrationManager.js. Move it from there to loaderHelper.js.
- Within loaderHelper, LoadPotree() is called when the document is ready.
  So actions within it do not need to wait further.
- I made the finishedLoading function, extracting the common code from the
  two callbacks for Potree.loadPointCloud().
